requirements.txt
Flask==0.10.1
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.2
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.9
WebHelpers==1.3
Werkzeug==0.10.1
aniso8601==0.92
gunicorn==19.3.0
itsdangerous==0.24
psycopg2==2.6
python-dateutil==2.4.1
pytz==2014.10
six==1.9.0
wsgiref==0.1.2
